+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This release adds the following:
 - The :proxied-auth option to most functions
 - The compare? function which invokes LDAP compare operation
+- get-connection and release-connection functions which promote proper use of the connection pool when a sequence of operations must be performed on a single connection
 
 ## [0.0.11]
 This release addresses a bug introduced in 0.0.10 and bumps the ldap-sdk version to 3.1.1.